refactor(core): move GlobalPerspective and components to core/perspective
