+++ packages/sanity/src/core/perspective/navbar/GlobalPerspectiveMenu.tsx
@@ -6,7 +6,7 @@ import {useCallback, useRef, useState} from 'react'
 import {styled} from 'styled-components'
 
 import {MenuButton} from '../../../ui-components'
-import {CreateReleaseDialog} from '../components/dialog/CreateReleaseDialog'
+import {CreateReleaseDialog} from '../../releases/components/dialog/CreateReleaseDialog'
 import {ReleasesList} from './ReleasesList'
 import {useScrollIndicatorVisibility} from './useScrollIndicatorVisibility'
