+++ packages/sanity/src/core/perspective/navbar/PerspectiveLayerIndicator.tsx
