+++ packages/sanity/src/core/perspective/navbar/ReleasesNav.tsx
@@ -9,10 +9,10 @@ import {useRouterState} from 'sanity/router'
 import {Tooltip} from '../../../ui-components'
 import {usePerspective} from '../../perspective/usePerspective'
 import {useSetPerspective} from '../../perspective/useSetPerspective'
+import {RELEASES_TOOL_NAME} from '../../releases/plugin'
+import {LATEST} from '../../releases/util/const'
+import {isDraftPerspective} from '../../releases/util/util'
 import {ToolLink} from '../../studio'
-import {RELEASES_TOOL_NAME} from '../plugin'
-import {LATEST} from '../util/const'
-import {isDraftPerspective} from '../util/util'
 import {CurrentGlobalPerspectiveLabel} from './currentGlobalPerspectiveLabel'
 import {GlobalPerspectiveMenu} from './GlobalPerspectiveMenu'
