+++ packages/sanity/src/core/perspective/navbar/useScrollIndicatorVisibility.ts
