+++ packages/sanity/src/core/releases/plugin/index.ts
@@ -1,8 +1,8 @@
 import {route} from 'sanity/router'
 
 import {definePlugin} from '../../config/definePlugin'
+import {ReleasesStudioNavbar} from '../../perspective/navbar/ReleasesStudioNavbar'
 import {releasesUsEnglishLocaleBundle} from '../i18n'
-import {ReleasesStudioNavbar} from '../navbar/ReleasesStudioNavbar'
 import {ReleasesTool} from '../tool/ReleasesTool'
 import resolveDocumentActions from './documentActions'
 import {ReleasesStudioLayout} from './ReleasesStudioLayout'
